mqtt2influx/requirements.txt
paho-mqtt
six
ujson
